backend/requirements.txt
Flask==3.0.0
PyYAML==6.0.1
Jinja2==3.1.2
requests==2.31.0
flask-cors==4.0.0
flasgger==0.9.7.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
cryptography==42.0.8